+++ src/modules/assignments/interfaces/assignment-service.interface.ts
@@ -4,5 +4,5 @@ import { AssignmentsWithRolesEntity } from '../entities'
 export interface IAssignmentService {
   databaseService: SumDatabaseService
 
-  getAssignmentsWithRolesByModelId(modelIds?: string[]): Promise<AssignmentsWithRolesEntity[]>
+  getAssignmentsWithRolesByModelId(modelIds?: string[], daysOfDelay?: number): Promise<AssignmentsWithRolesEntity[]>;
 }